From UINT_MAX to SIZE_MAX (or len)

+++ code/proof-methodologies/lemma-functions/ex-2-l_occurrences_of-props-answer.c
@@ -1,4 +1,4 @@
-#include <limits.h>
+#include <stdint.h>
 #include <stddef.h>
 
 /*@
@@ -118,7 +118,7 @@
         return i ;
       }
     }
-    return UINT_MAX;
+    return SIZE_MAX;
   }
 */
